make it properly and finished 2 & 3
-> more package to src folder
-> done with exercises 2 & 3
