--- prob2/EmployeeData.java
@@ -1,5 +1,0 @@
-package prob2;
-
-public interface EmployeeData {
-
-}
