--- prob2/Statistics.java
@@ -1,13 +1,0 @@
-package prob2;
-import java.util.*;
-public class Statistics {
-	/** 
-	 * Polymorphically computes and returns the sum
-	 * of all the salaries of all the staff/teachers in the list.
-	 */
-	public static double computeSumOfSalaries(List/*<implement>*/ aList) {
-		return 0.0;
-		//implement
-		//compute sum of all salaries of people in aList and return
-	}
-}
