+++ src/prob2/EmployeeData.java
@@ -1,0 +1,5 @@
+package prob2;
+
+public interface EmployeeData {
+    double getSalary();
+}
